requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
email-validator==2.2.0
apscheduler==3.10.4
pytz==2024.1
requests==2.31.0
stripe==10.0.0
python-dateutil==2.9.0
